predict/requirements.txt
opencv-python
numpy
Pillow
Flask==3.0.3
Werkzeug==3.0.3
flask-cors
requests